+++ src/CController.h
@@ -108,7 +108,7 @@ public:
     
     DissipationTracker* getDT();
 
-    
+    static ofstream _crosscheckdumpFilechem;
 };