+++ src/Structure/Compartment.h
@@ -125,11 +125,6 @@ protected:
 
 public:
     short _ID;
-/*    vector<uint16_t> cindex_bs;
-    vector<uint32_t> cID_bs;*/
-#ifdef MOTORBIASCHECK
-    size_t nummotorwalks=0;
-#endif
     cell_list::CellListHeadUser< Cylinder, Compartment > cylinderCell; // Cell of cylinders
     cell_list::CellListHeadUser< Triangle, Compartment > triangleCell; // Cell of triangles
     cell_list::CellListHeadUser< Edge,     Compartment > edgeCell;     // Cell of edges